Split into management and interaction
